Fix ci

+++ src/image.cpp
@@ -237,12 +237,14 @@ Image1u ReadImage1u(
             if (transform)
             {
                 ParallelFor(0, width * height, [&](int32 i) {
-                    image[i] = transform(std::max<uint8>(0, data[stride * i + channel]));
+                    image[i] = transform(std::max<uint8>(uint8(0), data[stride * i + channel]));
                 });
             }
             else
             {
-                ParallelFor(0, width * height, [&](int32 i) { image[i] = std::max<uint8>(0, data[stride * i + channel]); });
+                ParallelFor(0, width * height, [&](int32 i) {
+                    image[i] = std::max<uint8>(uint8(0), data[stride * i + channel]);
+                });
             }
         }
         else
@@ -251,14 +253,14 @@ Image1u ReadImage1u(
             {
                 for (int32 i = 0; i < width * height; ++i)
                 {
-                    image[i] = transform(std::max<uint8>(0, data[stride * i + channel]));
+                    image[i] = transform(std::max<uint8>(uint8(0), data[stride * i + channel]));
                 }
             }
             else
             {
                 for (int32 i = 0; i < width * height; ++i)
                 {
-                    image[i] = std::max<uint8>(0, data[stride * i + channel]);
+                    image[i] = std::max<uint8>(uint8(0), data[stride * i + channel]);
                 }
             }
         }
@@ -269,14 +271,14 @@ Image1u ReadImage1u(
         {
             for (int32 i = 0; i < width * height; ++i)
             {
-                image[i] = transform(Float(std::max<uint8>(0, data[stride * i + channel])));
+                image[i] = transform(Float(std::max<uint8>(uint8(0), data[stride * i + channel])));
             }
         }
         else
         {
             for (int32 i = 0; i < width * height; ++i)
             {
-                image[i] = Float(std::max<uint8>(0, data[stride * i + channel]));
+                image[i] = Float(std::max<uint8>(uint8(0), data[stride * i + channel]));
             }
         }
     }
